% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R, R/SelexPlots.R
\name{aopt}
\alias{aopt}
\title{aopt()}
\usage{
aopt(stock, nyears = 3, vbgf = NULL)

aopt(stock, nyears = 3, vbgf = NULL)
}
\arguments{
\item{stock}{class FLStock}
}
\value{
FLQuant with annual spr0y

FLQuant with annual spr0y
}
\description{
Function to compute Aopt, the age where an unfished cohort attains maximum biomass

Function to compute Aopt, the age where an unfished cohort attains maximum biomass
}
\author{
Henning Winker
}
