% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{brp.selex}
\alias{brp.selex}
\title{function to do nyears backtest of selex pattern in FLStocks}
\usage{
brp.selex(sel, stock, sr = NULL, Fref = NULL, nyears = 3)
}
\arguments{
\item{sel}{selex FLPars() or Sa FLQuants}

\item{stock}{stock object of class FLStock}

\item{sr}{spawner-recruitment function FLSR}

\item{Fref}{reference F for which Bref, Cref etc are computed (default=Fsq)}

\item{nyears}{number of years for reference conditions}
}
\value{
FLBRPs object
}
\description{
function to do nyears backtest of selex pattern in FLStocks
}
