% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelexPlots.R
\name{plotaopt}
\alias{plotaopt}
\title{plotaopt()}
\usage{
plotaopt(stock, nyears = 3)
}
\arguments{
\item{stock}{class FLStock}

\item{nyears}{number of years for biology}
}
\value{
ggplot
}
\description{
Function to compute Aopt, the age where an unfished cohort attains maximum biomass
}
\author{
Henning Winker
}
