% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelexPlots.R
\name{ploteqselex}
\alias{ploteqselex}
\title{ploteqselex()
return 2x2 plot showing F vs Sel for yield and ssb curves and isopleths}
\usage{
ploteqselex(brps, Fmax = 2, panels = NULL, ncol = 2, colours = NULL)
}
\arguments{
\item{brps}{output from brp.selex()}

\item{Fmax}{upper possible limit of  F range}

\item{panels}{choice of plots 1:4}

\item{ncol}{number of columns}

\item{colours}{optional, e.g. terrain.col, rainbow, etc.}
}
\value{
ggplot
}
\description{
ploteqselex()
return 2x2 plot showing F vs Sel for yield and ssb curves and isopleths
}
