% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelexPlots.R
\name{plotFselex}
\alias{plotFselex}
\title{plots trade-offs between relative Catch and SSB as a function Selectivity}
\usage{
plotFselex(brps, what = c("Fref", "Fmsy", "F0.1"), vbgf = NULL)
}
\arguments{
\item{brps}{output from brp.selex()}

\item{what}{type of F c("Fref","Fmsy","F0.1"), ref is by default Fcur}

\item{vbgf}{converts into length if von Bertalanffy FLPar is provided}
}
\value{
ggplot
}
\description{
plots trade-offs between relative Catch and SSB as a function Selectivity
}
