% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelexPlots.R
\name{plotprjselex}
\alias{plotprjselex}
\title{plotprjselex()
return 2x2 plot showing F vs Sel for yield and ssb curves and isopleths}
\usage{
plotprjselex(object, panels = NULL, ncol = NULL, colours = NULL, nyears = NULL)
}
\arguments{
\item{object}{FLStocks output from selex.forward/selex.backtest()}

\item{panels}{choice of plots 1:4}

\item{ncol}{number of columns}

\item{colours}{optional, e.g. terrain.col, rainbow, etc.}

\item{nyears}{number of years back from assessment year shown}
}
\value{
ggplot
}
\description{
plotprjselex()
return 2x2 plot showing F vs Sel for yield and ssb curves and isopleths
}
