% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelexPlots.R
\name{plotselex}
\alias{plotselex}
\title{plots mean selectivity at age Sa across selected years}
\usage{
plotselex(pars, Sa = NULL, obs = NULL, compounds = FALSE, colours = NULL)
}
\arguments{
\item{pars}{selexpars FLPars(s) or output from fitselex()}

\item{Sa}{observed selectivity-at-age (FLQuant) or FLStock}

\item{obs}{show observations if TRUE}

\item{compounds}{option to show selex compounds}
}
\value{
FLQuant of Selectivity-at-age Sa
}
\description{
plots mean selectivity at age Sa across selected years
}
