% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{selex.backtest}
\alias{selex.backtest}
\title{function to do nyears backtest of selex pattern in FLStocks}
\usage{
selex.backtest(sel, stock, sr = NULL, byears = 10, nyears = 3, plim = 1)
}
\arguments{
\item{sel}{list of selex parameters of class of FLPars()}

\item{stock}{stock object of class FLStock}

\item{sr}{optional spawner-recruitment function FLSR}

\item{byears}{number of backtest years}

\item{nyears}{number of years for reference conditions}

\item{plim}{set fbar for ages with Selectivy >= plim (default 0.975)}
}
\value{
FLStocks object
}
\description{
function to do nyears backtest of selex pattern in FLStocks
}
