% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{selex.fwd}
\alias{selex.fwd}
\title{function to forcast nyears under different selex patterns}
\usage{
selex.fwd(sel, stock, sr = NULL, fyears = 30, nyears = 3, plim = 1)
}
\arguments{
\item{sel}{list of selex parameters of class of FLPars()}

\item{stock}{stock object of class FLStock}

\item{sr}{optional spawner-recruitment function FLSR}

\item{fyears}{number of forecase years}

\item{nyears}{number of years for reference conditions}

\item{plim}{set fbar for ages with Selectivy >= plim (default 1)}

\item{Fref}{F to be projected, default Fsq}
}
\value{
FLStocks object
}
\description{
function to forcast nyears under different selex patterns
}
