% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{yopt.fwd}
\alias{yopt.fwd}
\title{function Extract yield and SSB}
\usage{
yopt.fwd(object, rel = FALSE, decimals = 3)
}
\arguments{
\item{object}{FLStocks output from selex.fwd()}

\item{rel}{if TRUE relative quantaties are shown}

\item{decimales}{}
}
\value{
Table with quatities
}
\description{
function Extract yield and SSB
}
