% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SelexPlots.R
\name{plotprjselex}
\alias{plotprjselex}
\title{plotprjselex()
return 2x2 plot showing F vs Sel for yield and ssb curves and isopleths}
\usage{
plotprjselex(
  object,
  panels = NULL,
  ncol = NULL,
  colours = NULL,
  nyears = NULL,
  max.discrete = 8
)
}
\arguments{
\item{object}{FLStocks output from selex.forward/selex.backtest()}

\item{ncol}{number of columns}

\item{colours}{optional, e.g. terrain.col, rainbow, etc.}

\item{nyears}{number of years back from assessment year shown}

\item{max.discrete}{determines max selectivities after which plot legend becomes continous}
}
\value{
ggplot
}
\description{
plotprjselex()
return 2x2 plot showing F vs Sel for yield and ssb curves and isopleths
}
