% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FLselex.R
\name{selex.backtest}
\alias{selex.backtest}
\title{function to do nyears backtest of selex pattern in FLStocks}
\usage{
selex.backtest(
  sel,
  stock,
  byears = 10,
  sr = NULL,
  nyears = 5,
  quantity = c("f", "catch")
)
}
\arguments{
\item{stock}{stock object of class FLStock}

\item{byears}{number of backtest years}

\item{sr}{spawner-recruitment function FLSR}

\item{nyears}{number of years for reference conditions}

\item{quantity}{observed "f" or "catch" for fwdControl()}

\item{pars}{list of selex parameters of class of FLPars()}
}
\value{
FLStocks object
}
\description{
function to do nyears backtest of selex pattern in FLStocks
}
